--- ruby-thrift/lib/app.rb
@@ -1,13 +1,0 @@
-module App
-  require_relative "app/consumer"
-
-  options = {
-    :url         => "amqp://localhost",
-    :exchange    => "ruby-exchange",
-    :routing_key => "ruby-messages",
-    :queue       => "ruby-queue"
-  }
-
-  Consumer.subscribe(options)
-
-end
