--- ruby-thrift/lib/app/publisher.rb
@@ -1,10 +1,0 @@
-module App
-  class Publisher
-    require "tackle"
-
-    def self.publish(message, options)
-      Tackle.publish(message, options)
-    end
-    
-  end
-end
